chore: Rename themes folder
