+++ .storybook/themeDecorator.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { ThemeProvider } from "styled-components";
-import standard from "../src/theme/standard";
+import standard from "../src/themes/standard";
 
 const ThemeDecorator = (storyFn) => (
   <ThemeProvider theme={standard}>{storyFn()}</ThemeProvider>
